split WlTouch out of wl_seat.cpp
