+++ src/server/frontend/wayland/wl_touch.h
@@ -1,0 +1,60 @@
+/*
+ * Copyright © 2018 Canonical Ltd.
+ *
+ * This program is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * Authored by: Christopher James Halse Rogers <christopher.halse.rogers@canonical.com>
+ */
+
+#ifndef WL_TOUCH_H
+#define WL_TOUCH_H
+
+#include "core_generated_interfaces.h"
+
+// from "mir_toolkit/events/event.h"
+struct MirInputEvent;
+
+namespace mir
+{
+
+class Executor;
+
+namespace frontend
+{
+
+class WlTouch : public wayland::Touch
+{
+public:
+    WlTouch(
+        wl_client* client,
+        wl_resource* parent,
+        uint32_t id,
+        std::function<void(WlTouch*)> const& on_destroy,
+        std::shared_ptr<mir::Executor> const& executor);
+
+    ~WlTouch();
+
+    void handle_event(MirInputEvent const* event, wl_resource* target);
+
+private:
+    std::shared_ptr<mir::Executor> const executor;
+    std::function<void(WlTouch*)> on_destroy;
+    std::shared_ptr<bool> const destroyed;
+
+    void release() override;
+};
+
+}
+}
+
+#endif // WL_TOUCH_H
